consolidate builder_txs
