+++ crates/builder/op-rbuilder/src/flashblocks/service.rs
@@ -10,12 +10,8 @@ use reth_payload_builder::{PayloadBuilderHandle, PayloadBuilderService};
 use reth_provider::CanonStateSubscriptions;
 
 use super::{
-    BuilderConfig, FlashblocksBuilderTx, FlashblocksNumberBuilderTx,
-    builder_tx::BuilderTransactions,
-    generator::BlockPayloadJobGenerator,
-    payload::{FlashblocksExecutionInfo, OpPayloadBuilder},
-    payload_handler::PayloadHandler,
-    wspub::WebSocketPublisher,
+    BuilderConfig, FlashblocksBuilderTx, generator::BlockPayloadJobGenerator,
+    payload::OpPayloadBuilder, payload_handler::PayloadHandler, wspub::WebSocketPublisher,
 };
 use crate::{
     metrics::OpRBuilderMetrics,
@@ -26,17 +22,15 @@ use crate::{
 pub struct FlashblocksServiceBuilder(pub BuilderConfig);
 
 impl FlashblocksServiceBuilder {
-    fn spawn_payload_builder_service<Node, Pool, BuilderTx>(
+    fn spawn_payload_builder_service<Node, Pool>(
         self,
         ctx: &BuilderContext<Node>,
         pool: Pool,
-        builder_tx: BuilderTx,
+        builder_tx: FlashblocksBuilderTx,
     ) -> eyre::Result<PayloadBuilderHandle<<Node::Types as NodeTypes>::Payload>>
     where
         Node: NodeBounds,
         Pool: PoolBounds,
-        BuilderTx:
-            BuilderTransactions<FlashblocksExecutionInfo> + Unpin + Clone + Send + Sync + 'static,
     {
         // TODO: is there a different global token?
         // this is effectively unused right now due to the usage of reth's `task_executor`.
@@ -110,22 +104,10 @@ where
         pool: Pool,
         _: OpEvmConfig,
     ) -> eyre::Result<PayloadBuilderHandle<<Node::Types as NodeTypes>::Payload>> {
-        let signer = self.0.builder_signer;
-
-        if let Some(builder_signer) = signer
-            && let Some(flashblocks_number_contract_address) =
-                self.0.flashblocks.flashblocks_number_contract_address
-        {
-            self.spawn_payload_builder_service(
-                ctx,
-                pool,
-                FlashblocksNumberBuilderTx::new(
-                    builder_signer,
-                    flashblocks_number_contract_address,
-                ),
-            )
-        } else {
-            self.spawn_payload_builder_service(ctx, pool, FlashblocksBuilderTx::new(signer))
-        }
+        let builder_tx = FlashblocksBuilderTx::new(
+            self.0.builder_signer,
+            self.0.flashblocks.flashblocks_number_contract_address,
+        );
+        self.spawn_payload_builder_service(ctx, pool, builder_tx)
     }
 }
