+++ crates/shared/jwt/Cargo.toml
@@ -12,9 +12,7 @@ repository.workspace = true
 workspace = true
 
 [features]
-test-utils = [
-	"kona-engine?/test-utils"
-]
+test-utils = [ "kona-engine?/test-utils" ]
 engine-validation = [
 	"dep:alloy-provider",
 	"dep:alloy-transport-http",
